Refactor code for simpler use
